Migrate Oled to keyboard folder
