+++ keyboards/crkbd/keymaps/mb_via/config.h
@@ -31,10 +31,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #    undef RGBLED_SPLIT
 #    define RGBLED_SPLIT \
         { 27, 27 }
-#    define RGBLIGHT_LIMIT_VAL 120
 #    define RGBLIGHT_HUE_STEP  10
 #    define RGBLIGHT_SAT_STEP  17
 #    define RGBLIGHT_VAL_STEP  17
 #endif
-
-#define OLED_FONT_H "keyboards/crkbd/lib/glcdfont.c"
